requirements.txt
openai-whisper
moviepy
torch
torchaudio
numpy
ffmpeg-python
yt-dlp
